fix: resolve root compilation errors (getString, MutablePreferences, imageLoader, type mismatches, Composable annotations)

+++ app/src/main/kotlin/com/exapps/mangaworld/core/firebase/ChapterUpdateChecker.kt
@@ -110,7 +110,7 @@ class ChapterUpdateChecker @Inject constructor(
                                 coverUrl = favorite.coverUrl
                             )
                         )
-                        prefs.edit().putFloat("max_chapter_${favorite.mangaId}", currentMaxChapter).apply()
+                        prefs.edit().putFloat("max_chapter_${favorite.mangaId}", currentMaxChapter.toFloat()).apply()
                     }
                 }
             } catch (e: Exception) {
